[jets] move jets_math under jets
