+++ rust/ares/src/lib.rs
@@ -2,7 +2,6 @@
 extern crate num_derive;
 pub mod interpreter;
 pub mod jets;
-pub mod jets_math;
 pub mod mem;
 pub mod mug;
 pub mod newt;
